requirements.txt
boto==2.48.0
certifi==2018.1.18
chardet==3.0.4
Collectfast==0.6.0
Django==2.0.2
django-appconf==1.0.2
django-ckeditor==5.4.0
django-embed-video==1.1.2
django-environ==0.4.4
django-gunicorn==0.1.1
django-imagekit==4.0.2
django-js-asset==1.0.0
django-solo==1.1.3
django-storages==1.6.5
greenlet==0.4.13
gunicorn==19.7.1
idna==2.6
msgpack==0.5.4
msgpack-python==0.5.4
neovim==0.2.1
olefile==0.45.1
pilkit==2.0
Pillow==5.0.0
psycopg2-binary==2.7.4
python-dateutil==2.6.1
pytz==2018.3
requests==2.18.4
six==1.11.0
South==1.0.2
static3==0.7.0
urllib3==1.22
